requirements.txt
boto3==1.34.69
python-dotenv==1.0.1
pandas==2.2.1  # Para formatação de dados 
flask==3.0.2
psutil==5.9.8 